domain shutting down added
