+++ api/conn/connect.go
@@ -62,8 +62,6 @@ func Use(vals ...interface{}){
 	for _,val:= range vals{
 	 _=val
 	}
-	
-
 }
