+++ server/server.go
@@ -13,7 +13,8 @@ func InitServer(portNum int, libvirtInst *conn.InstHandler, domain *conn.Domain)
 	
 	http.HandleFunc("GET /getStatus", libvirtInst.ReturnStatus)
 	http.HandleFunc("POST /createVM", libvirtInst.CreateVM)
-
+	http.HandleFunc("Get /getStatusUUID", libvirtInst.ReturnStatusUUID)
+	http.HandleFunc("Get /ForceShutDownUUID", libvirtInst.ForceShutDownVM)
 
 	log.Fatal(http.ListenAndServe(":"+strconv.Itoa(portNum), nil))
 }
